move shared cache credential to its own module
