+++ sdk/identity/azure-identity/azure/identity/_credentials/default.py
@@ -12,7 +12,7 @@ from .browser import InteractiveBrowserCredential
 from .chained import ChainedTokenCredential
 from .environment import EnvironmentCredential
 from .managed_identity import ManagedIdentityCredential
-from .user import SharedTokenCacheCredential
+from .shared_cache import SharedTokenCacheCredential
 
 _LOGGER = logging.getLogger(__name__)
